example progress

+++ example/TablePrinter.cpp
@@ -57,7 +57,7 @@ void TablePrinter::print_element(std::ostream &out, const oos::object_base_ptr &
   out << line_.str() << "\n";
 }
 
-unsigned int TablePrinter::column_width(const char *id, unsigned long min) const
+unsigned int TablePrinter::column_width(const char *id, size_t min) const
 {
   return 2 + std::max(strlen(id), min) + 1;
 }
